fix(vendors): move apply schema to non-'use server' module
Next.js 16 rejects const/type exports from 'use server' files — moved
vendorApplicationSchema + types to apply-schema.ts so apply.ts only
exports async server actions. Fixes the prod build.
