+++ src/domains/vendors/index.ts
@@ -1,5 +1,6 @@
 export * from './actions'
 export * from './apply'
+export * from './apply-schema'
 export * from './brand-claims'
 export * from './onboarding'
 export * from './stripe'
